improve planning efficiency (#196)
* Add efficient planning changes

* fixing dtype handling in solvers

---------

+++ stable_worldmodel/policy.py
@@ -468,7 +468,7 @@ class WorldModelPolicy(BasePolicy):
                 action[i] = self._action_buffer[i].popleft()
 
         action = action.reshape(*self.env.action_space.shape)
-        action = action.numpy()
+        action = action.float().numpy()
 
         if 'action' in self.process:
             action = self.process['action'].inverse_transform(action)
